more junie cleanup
