+++ krrood/src/krrood/entity_query_language/cache_data.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from krrood.entity_query_language.utils import T, ensure_hashable
+from krrood.utils import T, ensure_hashable
 
 """
 Cache utilities.
